Refactoring to Deposit workflow in preparation of second flow
